HSEARCH-4235 Add a dedicated test profile for ES 7.12+
