+++ CONTRIBUTING.md
@@ -207,7 +207,8 @@ The following profiles are available:
   * `elasticsearch-7.7` for 7.7
   * `elasticsearch-7.8` for 7.8 to 7.9
   * `elasticsearch-7.10` for 7.10 (**the default**)
-  * `elasticsearch-7.11` for 7.11+ ([not open-source](https://opensource.org/node/1099))
+  * `elasticsearch-7.11` for 7.11 ([not open-source](https://opensource.org/node/1099))
+  * `elasticsearch-7.12` for 7.12+ ([not open-source](https://opensource.org/node/1099))
 * [Open Distro for Elasticsearch](https://opendistro.github.io/for-elasticsearch/):
   * `opendistro-elasticsearch-1.13` for 1.13+
