+++ pom.xml
@@ -215,6 +215,7 @@
         <!-- The latest micro of each Elasticsearch branch -->
         <version.org.elasticsearch.latest-7.13>7.13.2</version.org.elasticsearch.latest-7.13>
         <version.org.elasticsearch.latest-7.12>7.12.1</version.org.elasticsearch.latest-7.12>
+        <version.org.elasticsearch.latest-7.11>7.11.2</version.org.elasticsearch.latest-7.11>
         <version.org.elasticsearch.latest-7.10>7.10.0</version.org.elasticsearch.latest-7.10>
         <version.org.elasticsearch.latest-7.9>7.9.3</version.org.elasticsearch.latest-7.9>
         <version.org.elasticsearch.latest-7.7>7.7.1</version.org.elasticsearch.latest-7.7>
